finalized browserify file structure
