+++ app/index.html
@@ -8,9 +8,6 @@
 
       <link rel="stylesheet" type="text/css" href="/build/css/style.css">
 
-      <!-- build:js build/js/vendor/modernizr.js -->
-      <!-- <script src="/bower_components/modernizr/modernizr.js"></script> -->
-      <!-- endbuild -->
   </head>
   <body>
 
@@ -24,71 +21,7 @@
       </section>
     </section>
 
+    <script type="text/javascript" src="build/js/app.build.js"></script>
 
-    <!-- Templates -->
-    <!-- ===================================== -->
-
-    <!-- Single Project Template -->
-    <!-- ===================================== -->
-    <script type="text/template" id="single-project-template">
-      <h2><%= title %></h2>
-
-      <section id="sections">
-        <input id="new-section"><button id="add-new-section">Add New</button>
-        <div id="section-list">
-        </div>
-      </section>
-    </script>
-
-    <!-- Single Section Template -->
-    <!-- ===================================== -->
-    <!-- <script type="text/template" id="single-section-template">
-      <h2><%- title %></h2>
-      <section id="versions">
-        <input id="new-version" type="file"><button id="add-new-version">Add New</button>
-        <div id="version-list">
-        </div>
-      </section>
-    </script> -->
-
-    <!-- Project Template -->
-    <!-- ===================================== -->
-    <script type="text/template" id="project-template">
-      <a href="#" id="open-project">
-        <div class="single-project"><h2><%- title %></h3><br><a href="#" id="delete-project">Delete</a></div>
-      </a>
-    </script>
-
-    <!-- Section Template -->
-    <!-- ===================================== -->
-    <script type="text/template" id="section-template">
-      <a href="#" id="open-section">
-        <div class="single-section">
-          <h2><%= title %></h2>
-          <br>
-          <a href="#" id="delete-section">Delete</a>
-        </div>
-      </a>
-    </script>
-
-    <!-- Version Template -->
-    <!-- ===================================== -->
-    <script type="text/template" id="version-template">
-      <div class="single-version"><br><a href="#" id="delete-version">Delete</a></div>
-    </script>
-
-    <!-- bower:js -->
-    <script src="/bower_components/jquery/dist/jquery.js"></script>
-    <script src="/bower_components/underscore/underscore.js"></script>
-    <script src="/bower_components/backbone/backbone.js"></script>
-    <!-- endbower -->
-
-    <!-- build:js build/js/build.app.js -->
-    <script type="text/javascript" src="/js/backbone-model-file-upload.js"></script>
-    <script type="text/javascript" src="/js/models/models.js"></script>
-    <script type="text/javascript" src="/js/collections/collections.js"></script>
-    <script type="text/javascript" src="/js/views/views.js"></script>
-    <script type="text/javascript" src="/js/app.js"></script>
-    <!-- endbuild -->
   </body>
 </html>
