+++ app/js/collections/project-list.js
@@ -1,5 +1,7 @@
 var Backbone = require('backbone');
 
+var Project = require('../models/project');
+
 module.exports = Backbone.Collection.extend({
   model: Project,
   url: 'http://totem-backend12.herokuapp.com/projects'
