+++ app/js/collections/section-list.js
@@ -1,5 +1,7 @@
 var Backbone = require('backbone');
 
+var Section = require('../models/section');
+
 module.exports = Backbone.Collection.extend({
   model: Section,
   url: 'http://totem-backend12.herokuapp.com/sections'
