--- app/js/models/models.js
@@ -1,26 +1,0 @@
-var Backbone = require('backbone');
-
-Project = Backbone.Model.extend({
-	urlRoot: 'http://totem-backend12.herokuapp.com/projects',
-	defaults: {
-	    title: ''
-	}
-});
-
-Section = Backbone.Model.extend({
-	urlRoot: 'http://totem-backend12.herokuapp.com/sections',
-	defaults: {
-	    title: '',
-	    project_id: ''
-	}
-});
-
-Version = Backbone.Model.extend({
-  urlRoot: 'http://totem-backend12.herokuapp.com/versions',
-  defaults: {
-      section_id: '',
-      file: ''
-  }
-});
-
-
