+++ app/js/templates/project.html
@@ -1,0 +1,7 @@
+<a href="#" id="open-project">
+  <div class="single-project">
+    <h2><%- title %></h2>
+    <br>
+    <a href="#" id="delete-project">Delete</a>
+  </div>
+</a>
