+++ app/js/templates/section.html
@@ -1,0 +1,7 @@
+<a href="#" id="open-section">
+  <div class="single-section">
+    <h2><%= title %></h2>
+    <br>
+    <a href="#" id="delete-section">Delete</a>
+  </div>
+</a>
