+++ app/js/templates/single-project.html
@@ -1,0 +1,7 @@
+<h2><%= title %></h2>
+
+<section id="sections">
+  <input id="new-section">
+  <button id="add-new-section">Add New</button>
+  <div id="section-list"></div>
+</section>
