+++ app/js/templates/version.html
@@ -1,0 +1,4 @@
+<div class="single-version">
+  <br>
+  <a href="#" id="delete-version">Delete</a>
+</div>
